requirements.txt
chess~=1.11.2
datasets~=3.5.0
evaluate~=0.4.3
jiwer~=3.1.0
matplotlib~=3.10.1
mltu~=1.2.5
numpy==1.26.4
opencv_python
pandas~=2.2.3
PyYAML~=6.0.2
tiktoken~=0.9.0
torch==2.2.0
torchsummaryX~=1.3.0
tqdm~=4.67.1
transformers~=4.50.3
onnx~=1.17.0
sentencepiece~=0.2.0
tensorboard~=2.19.0
tiktoken
timm~=1.0.15

opencv-python~=4.11.0.86
torchvision~=0.17.0
pillow~=11.1.0
defusedxml~=0.7.1
ipython~=9.0.2
beautifulsoup4~=4.13.3
soupsieve~=2.6
packaging~=24.2
h11~=0.14.0
jupyter~=1.1.1
pip~=24.0
attrs~=25.3.0
wheel~=0.37.1
tornado~=6.4.2
jinja2~=3.1.6
filelock~=3.18.0
pyzmq~=26.4.0
pexpect~=4.9.0
cffi~=1.17.1
setuptools~=78.1.0
comm~=0.2.2
traitlets~=5.14.3
dill~=0.3.8
fqdn~=1.5.1
parso~=0.8.4
jedi~=0.19.2
qqdm~=0.0.7
onnxruntime~=1.21.0
pytz~=2025.2
addict~=2.4.0
ipywidgets~=8.1.5
safetensors~=0.5.3
requests~=2.32.3
notebook~=7.3.3
idna~=3.10
multidict~=6.3.2
propcache~=0.3.1
sniffio~=1.3.1
arrow~=1.3.0
python-dateutil~=2.9.0.post0
babel~=2.17.0
click~=8.1.8
pygments~=2.19.1
httpcore~=1.0.7
httpx~=0.28.1
rapidfuzz~=3.13.0
json5~=0.12.0
overrides~=7.7.0
regex~=2024.11.6
sympy~=1.13.3
mpmath~=1.3.0
decorator~=5.2.1
triton~=2.2.0
psutil~=7.0.0
urllib3~=2.3.0
fsspec~=2024.12.0
bleach~=6.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
aiohttp~=3.11.16
yarl~=1.18.3
MarkupSafe~=3.0.2
pyarrow~=19.0.1
wcwidth~=0.2.13
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.6.1
certifi~=2025.1.31
debugpy~=1.8.13
Markdown~=3.7
ptyprocess~=0.7.0
tzdata~=2025.2
multiprocess~=0.70.16
xxhash~=3.5.0
six~=1.17.0
anyio~=4.9.0
nbformat~=5.10.4
nbclient~=0.10.2
fastjsonschema~=2.21.1
jsonschema~=4.23.0
networkx~=3.4.2
jupyterlab~=4.3.6
asttokens~=3.0.0
contourpy~=1.3.1
executing~=2.2.0
fonttools~=4.57.0
ipykernel~=6.29.5
nbconvert~=7.16.6
pandocfilters~=1.5.1
mistune~=3.1.3
pycparser~=2.22
pyparsing~=3.2.3
websocket-client~=1.8.0
terminado~=0.18.1
referencing~=0.36.2
webcolors~=24.11.1
jsonpointer~=3.0.0
isoduration~=20.11.0
cycler~=0.12.1
kiwisolver~=1.4.8
Send2Trash~=1.8.3
tomli~=2.0.1
autocommand~=2.2.2
inflect~=7.3.1
typeguard~=4.3.0
platformdirs~=4.3.7
zipp~=3.19.2
tokenizers~=0.21.1
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=25.2.10
werkzeug~=3.1.3